Director: Separated definition and implementation, updating namespace
